feat: divide code into modules, delete function.js, make DESCRIPTION_PHOTO plural
